#33 Move cube accessor to cube module
