+++ decode/core/__init__.py
@@ -30,4 +30,3 @@ class BaseAccessor(object):
 from .array.classes import *
 from .array.decorators import *
 from .array.functions import *
-from .cube.classes import *
